+++ src/SzamlaAgent.php
@@ -54,11 +54,15 @@ class SzamlaAgent
 
     private bool $singleton = true;
 
-    protected function __construct(?string $username, ?string $password, ?string $apiKey, bool $downloadPdf, int $responseType = AbstractResponse::RESULT_AS_XML, string $aggregator = '')
+    protected function __construct(?string $username, ?string $password, ?string $apiKey, ?bool $downloadPdf, int $responseType = AbstractResponse::RESULT_AS_XML, string $aggregator = '')
     {
         $this->setting = new SzamlaAgentSetting($username, $password, $apiKey, $downloadPdf, SzamlaAgentSetting::DOWNLOAD_COPIES_COUNT, $responseType, $aggregator);
         $this->cookieHandler = new CookieHandler();
         Log::channel('szamlazzhu')->debug(sprintf('Számla Agent initialization is complete ($username: %s, apiKey: %s)', $username, $apiKey));
+
+        if (null === $downloadPdf) {
+            $downloadPdf = config('szamlazzhu.pdf.file_save', false);
+        }
 
         $this->isPdfFileSaveable = $downloadPdf;
         $this->isXmlFileSaveable = config('szamlazzhu.xml.file_save', false);
@@ -89,7 +93,7 @@ class SzamlaAgent
     /**
      * API key is the recommended authentication mode
      */
-    public static function createWithAPIkey(string $apiKey, bool $downloadPdf = true, int $responseType = AbstractResponse::RESULT_AS_XML, string $aggregator = '')
+    public static function createWithAPIkey(string $apiKey, ?bool $downloadPdf = null, int $responseType = AbstractResponse::RESULT_AS_XML, string $aggregator = '')
     {
         $index = self::getHash($apiKey);
 
@@ -137,11 +141,16 @@ class SzamlaAgent
     private function sendRequest(SzamlaAgentRequest $request): AbstractResponse
     {
         $this->setRequest($request);
-        if (Proforma::class === $request->getEntity()::class) {
-            return new ProformaDeletionResponse($this, $request->send());
-        } elseif (Invoice::class === $request->getEntity()::class) {
+        $entity = $request->getEntity();
+        if ($entity instanceof Proforma) {
+            if ($request->getType() === 'deleteProforma') {
+                return new ProformaDeletionResponse($this, $request->send());
+            }
+
             return new InvoiceResponse($this, $request->send());
-        } elseif (Receipt::class === $request->getEntity()::class) {
+        } elseif ($entity instanceof Invoice) {
+            return new InvoiceResponse($this, $request->send());
+        } elseif ($entity instanceof Receipt) {
             return new ReceiptResponse($this, $request->send());
         }
 
@@ -653,6 +662,11 @@ class SzamlaAgent
     public function setResponseXmlFileSave(bool $isResponseXmlFileSaveable): void
     {
         $this->isResponseXmlFileSaveable = $isResponseXmlFileSaveable;
+    }
+
+    public function setPdfFileSaveable(bool $isPdfFileSaveable): void
+    {
+        $this->isPdfFileSaveable = $isPdfFileSaveable;
     }
 
     /**